+++ CHANGES.rst
@@ -4,4 +4,6 @@ Changelog
 1.0.0
 -----
 
+- #2 Support for custom logos in results report for client and lab
+- #2 Added default view controller for results reports
 - #1 Add Zendesk's BES Support Centre Icon